Move leadership to its own package
